server/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-chess==1.999
